Seperate getRawQuery and getExpiration from /web-app/pages/api/cert/[imageFileName].ts into its own file at /web-app/helpers/expiration-query.ts Refactor [imageFileName].ts to import getExpiration from expiration-query.ts Refactor Synth data generator to generate data from data sources: - /web-app/synth/account_names.json - /web-app/synth/receipt_ids.json - /web-app/synth/timestamps.json Synth data schemas: - /web-app/synth/test_db/action_receipts.json - /web-app/synth/test_db/receipts.json Remove beforeAll() function in /web-app/__tests__/test-query.test.ts since Synth will take care of populaating the test db.
