+++ web-app/__tests__/test-query.test.ts
@@ -1,75 +1,8 @@
 import { PrismaClient } from '@prisma/client';
-import getExpiration from '../prisma/test-helpers/expiration';
+import getExpiration from '../helpers/expiration-query';
 import { getQueryResult } from '../prisma/test-helpers/query';
 
 const prisma = new PrismaClient();
-
-// eslint-disable-next-line max-lines-per-function
-beforeAll(async () => {
-  // create receipts
-  await prisma.receipts.createMany({
-    data: [
-      {
-        receipt_id: 'xxxxxxxxxxxxxxxxsxxxxxxxxxxxxxxxsxxxxxxxxxxxx',
-        included_in_block_timestamp: 1646224546000000000,
-      },
-      {
-        receipt_id: 'yyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy',
-        included_in_block_timestamp: 1745788799000000000,
-      },
-      {
-        receipt_id: 'zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz',
-        included_in_block_timestamp: 1671788799000000000,
-      },
-      {
-        receipt_id: 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
-        included_in_block_timestamp: 1646224546000000000,
-      },
-      {
-        receipt_id: 'bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
-        included_in_block_timestamp: 1647224546000000000,
-      },
-      {
-        receipt_id: 'ccccccccccccccccccccccccccccccccccccccccccccc',
-        included_in_block_timestamp: 1648224546000000000,
-      },
-    ],
-  });
-
-  // create action_receipts
-  await prisma.action_receipts.createMany({
-    data: [
-      {
-        receipt_id: 'xxxxxxxxxxxxxxxxsxxxxxxxxxxxxxxxsxxxxxxxxxxxx',
-        signer_account_id: 'janedoe.testnet',
-      },
-      {
-        receipt_id: 'yyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy',
-        signer_account_id: 'janedoe.testnet',
-      },
-      {
-        receipt_id: 'zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz',
-        signer_account_id: 'janedoe.testnet',
-      },
-      {
-        receipt_id: 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
-        signer_account_id: 'johndoe.testnet',
-      },
-      {
-        receipt_id: 'bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
-        signer_account_id: 'johndoe.testnet',
-      },
-      {
-        receipt_id: 'ccccccccccccccccccccccccccccccccccccccccccccc',
-        signer_account_id: 'johndoe.testnet',
-      },
-    ],
-  });
-
-  console.log('✨ 2 action_receipts successfully created!');
-
-  console.log('✨ 2 receipts successfully created!');
-});
 
 afterAll(async () => {
   const deleteReceipt = prisma.action_receipts.deleteMany();
@@ -79,7 +12,7 @@ afterAll(async () => {
 
   await prisma.$disconnect();
 });
-
+// TODO: REFACTOR
 it('should return expiration date for jane doe', async () => {
   // jane has a 180-dayinactivity after issue date
   await expect(getExpiration('janedoe.testnet', '2022-03-02')).resolves.toEqual('2022-08-29');
