+++ web-app/synth/account_names.json
@@ -1,0 +1,1 @@
+["janedoe.testnet", "johndoe.testnet"]
