+++ web-app/synth/receipt_ids.json
@@ -1,0 +1,8 @@
+[
+  "r74mcWqH2zCQeBzOgiS4skLnjvARIONorhfKroxrFAEts",
+  "o2uftO97TSisawK2PCkqe45tki0A9IX6lWRkQXUjy6rW6",
+  "Zc82R6f58evLaZ3h306k9vs9PpAifXytsRABt4ngpHa6V",
+  "0zxA9dIZCrd8NltfLA3xFy2ctNTT4nskvNkjZP5bvEBl5",
+  "Wt4a5NwKgihcWiKlU6NHDWhfoeE9b7HsYUIjTQAfCUoic",
+  "OkcSMfikcRDP1xGRiRMSVPMciC2Mq1tndRC2Mq1tndRC2"
+]
