--- web-app/synth/receipts.json
@@ -1,25 +1,0 @@
-{
-  "type": "array",
-  "length": 3,
-  "content": {
-    "type": "object",
-    "receipt_id": {
-      "type": "string",
-      "truncated": {
-        "content": {
-          "type": "string",
-          "pattern": "[a-zA-Z0-9]{0, 255}"
-        },
-        "length": 45
-      }
-    },
-    "included_in_block_timestamp": {
-      "type": "number",
-      "categorical": {
-        "1646224546000000000": 1,
-        "1745788799000000000": 1,
-        "1671788799000000000": 1
-      }
-    }
-  }
-}
