+++ web-app/synth/test_db/action_receipts.json
@@ -1,0 +1,13 @@
+{
+  "type": "array",
+  "length": 6,
+  "content": {
+    "type": "object",
+    "receipt_id": "@receipts.content.receipt_id",
+    "signer_account_id": {
+      "type": "datasource",
+      "path": "json:synth/account_names.json",
+      "cycle": true
+    }
+  }
+}
