+++ web-app/synth/test_db/receipts.json
@@ -1,0 +1,17 @@
+{
+  "type": "array",
+  "length": 6,
+  "content": {
+    "type": "object",
+    "receipt_id": {
+      "type": "datasource",
+      "path": "json:synth/receipt_ids.json",
+      "cycle": false
+    },
+    "included_in_block_timestamp": {
+      "type": "datasource",
+      "path": "json:synth/timestamps.json",
+      "cycle": false
+    }
+  }
+}
